separate out admin routes to own router and namespace
all admin pages behind /admin
