+++ public/login_page.html
@@ -16,7 +16,7 @@
   <h1 style = "color:#4D4949;" class = "center"> <font size = "6" > Ears for Peers Admin Login </font></h1>
   
 
-  <form action='/login', method='post'>
+  <form action='/admin/login', method='post'>
     <p>
     <input class = "input_box"
            type = "username"
