+++ public/register_page.html
@@ -15,7 +15,7 @@
   <h1 style = "color:#4D4949;" class = "center"> <font size = "6" > Ears for Peers Admin Registration </font></h1>
   
 
-  <form action='/register', method='post'>
+  <form action='/admin/register', method='post'>
     <p>
     <input class = "input_box"
            type = "username"
