+++ routes/adminRoutes.js
@@ -1,18 +1,57 @@
-// TODO abstract out admin router
-
 const express = require('express');
 const passport = require('passport');
+const path = require('path');
+
 let router = express.Router();
 let Admin = require('../models/adminModel');
 
-router.get('/login', () => {});
+function ensureAuthenticated(req, res, next) {
+  if (process.env.NOAUTH) { return next(); }
+  if (req.isAuthenticated()) { return next(); }
+  res.redirect('/admin/login');
+}
 
-router.post('/login', () => {});
+router.get('/', ensureAuthenticated, (req, res) => {
+  // console.log("GET /admin")
+  res.sendFile('admin.html', {root: path.join(__dirname, '../public')});
+});
 
-router.get('/logout', () => {});
+router.get('/login', (req, res) => {
+  // console.log("GET /admin/login")
+  res.sendFile('login_page.html', {root: path.join(__dirname, '../public')});
+});
 
-router.get('/register', () => {});
+router.post('/login', passport.authenticate('local', { failureRedirect: '/admin/login' }), (req, res) => {
+  // console.log("POST /admin/login")
+  res.redirect('/admin');
+});
 
-router.post('/register', () => {});
+router.get('/logout', ensureAuthenticated, (req, res) => {
+  // console.log("GET /admin/logout")
+  req.logout();
+  res.redirect('/admin/login');
+});
+
+router.get('/register', ensureAuthenticated, (req, res) => {
+  // console.log("GET /admin/register")
+  res.sendFile('register_page.html', {root: path.join(__dirname, '../public')});
+});
+
+router.post('/register', ensureAuthenticated, (req, res) => {
+  // console.log("POST /admin/register")
+  Admin.register(new Admin({username : req.body.username }), req.body.password, function(err, admin) {
+    if (err) {
+      console.log(err);
+      res.sendStatus(500);
+    } else {
+      res.redirect('/admin/login');
+    }
+  });
+});
+
+// TODO THIS IS NOT SECURE, PROTECT ALL ADMIN ROUTES BEHIND AUTH
+router.get('/:folder/:file', function(req, res) {
+  res.sendFile(req.params.file, {root: path.join(__dirname, '../public', req.params.folder)});
+});
 
 module.exports = router;
